08_Project/06_E_Commerce_Chatbot_main/requirements.txt
langchain-astradb
langchain 
# langchain-openai
datasets
pypdf
python-dotenv
flask

#If we use google gemin embedding
google-generativeai
langchain-community
langchain-google-genai


-e .